pyproject.toml
[tool.poetry]
name = "flux-backend"
version = "0.1.0"
description = ""
authors = ["Jumbik_Tank <oprivetdaent@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.114.0"
pydantic = ">=1.8.2"
SQLAlchemy = ">=2.0.34"
uvicorn = ">=0.15.0"
python-dotenv = "1.0.1"
uuid6 = "^2024.7.10"
pandas = "^2.2.2"
flux-orm = {git = "ssh://git@github.com/FluxFury/flux-orm.git", rev = "main"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
